@@ -1,61 +1,99 @@
 plugins {
-    id 'com.github.johnrengelman.shadow' version '5.0.0'
-    id "io.freefair.lombok" version "3.7.5"
-    id 'java'
+    id 'com.github.johnrengelman.shadow' version '7.1.2'
+    id 'io.freefair.lombok' version '6.3.0'
+    id 'java-library'
+    id 'org.sonarqube' version '3.3'
 }
 
 group = 'me.tychsen'
-version  = '1.5.12'
-
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
+version  = '1.7.0'
 
 repositories {
     mavenCentral()
     maven {
-        name= "aikar-acf"
-        url = "https://repo.aikar.co/content/groups/aikar/" }
+        name= 'aikar-acf'
+        url = 'https://repo.aikar.co/content/groups/aikar/' }
     maven {
-        name = "spigot-repo"
-        url = "https://hub.spigotmc.org/nexus/content/repositories/snapshots/"
+        name = 'spigot-repo'
+        url = 'https://hub.spigotmc.org/nexus/content/repositories/snapshots/'
     }
     maven {
-        name = "vault-repo"
-        url = "http://nexus.hc.to/content/repositories/pub_releases"
+        name = 'vault-repo'
+        url = 'https://nexus.hc.to/content/repositories/pub_releases'
     }
     maven {
-        name = "bungeecord-chat"
-        url = "https://oss.sonatype.org/content/groups/public/"
+        name = 'bungeecord-chat'
+        url = 'https://oss.sonatype.org/content/groups/public/'
     }
     maven {
-        name = "codemc-repo"
-        url = "https://repo.codemc.org/repository/maven-public"
+        name = 'codemc-repo'
+        url = 'https://repo.codemc.org/repository/maven-public'
+    }
+    maven {
+        name = 'jitpack'
+        url = 'https://jitpack.io'
+    }
+    maven {
+        name = 'playerpoints'
+        url = 'https://repo.rosewooddev.io/repository/public/'
+    }
+    maven {
+        name = "triumph-gui"
+        url = "https://repo.mattstudios.me/artifactory/public/"
     }
 }
 
 dependencies {
-    implementation "org.spigotmc:spigot-api:1.16.1-R0.1-SNAPSHOT"
-    implementation "net.milkbowl.vault:VaultAPI:1.7"
-    compile "co.aikar:acf-bukkit:0.5.0-SNAPSHOT"
-    compile "org.bstats:bstats-bukkit:1.7"
-    compile group: 'org.jetbrains', name: 'annotations', version: '19.0.0'
+    compileOnly 'org.spigotmc:spigot-api:1.17-R0.1-SNAPSHOT'
+    compileOnly 'com.github.MilkBowl:VaultAPI:1.7.1'
+    compileOnly 'org.black_ixx:playerpoints:3.2.3'
+    api "dev.triumphteam:triumph-gui:3.1.2"
+    api 'com.github.sarhatabaot:krakencore:1.1.3'
+    api 'de.tr7zw:item-nbt-api:2.9.2'
+    api 'co.aikar:acf-paper:0.5.1-SNAPSHOT'
+    api 'org.bstats:bstats-bukkit:3.0.0'
+    api group: 'org.jetbrains', name: 'annotations', version: '23.0.0'
 }
 
-import org.apache.tools.ant.filters.ReplaceTokens
+rootProject.tasks.named("processResources") {
+    duplicatesStrategy = DuplicatesStrategy.INCLUDE
+}
+
 
 processResources {
-    from(sourceSets.main.resources.srcDirs) {
-        filter ReplaceTokens, tokens: [version: version]
+    def props = [version: version]
+    inputs.properties props
+    filteringCharset 'UTF-8'
+    filesMatching('plugin.yml') {
+        expand props
     }
 }
 
 shadowJar {
-    classifier = ""
-    configurations = [project.configurations.compile]
+    archiveClassifier.set("")
     relocate 'org.bstats', 'me.tychsen.enchantgui.util'
+    relocate 'co.aikar.commands', 'me.tychsen.enchantgui.acf'
+    relocate 'co.aikar.locales', 'me.tychsen.enchantgui.locales'
+    relocate 'de.tr7zw', 'me.tychsen.enchantgui.nbt'
+    relocate 'dev.triumphteam.gui', 'me.tychsen.enchantguigui'
+    relocate 'com.github.sarhatabaot.kraken', 'me.tychsen.enchantgui.kraken'
 }
 
 wrapper {
-    gradleVersion = "5.4.1"
-    
+    gradleVersion = '7.4.2'
 }
+
+sonarqube {
+    properties {
+        property 'sonar.projectKey', 'sarhatabaot_EnchantGUI'
+        property 'sonar.organization', 'sarhatabaot-github'
+        property 'sonar.host.url', 'https://sonarcloud.io'
+    }
+}
+
+java {
+    toolchain {
+        languageVersion.set(JavaLanguageVersion.of(16))
+    }
+}
+
