@@ -1,93 +1,80 @@
 package me.tychsen.enchantgui.config;
 
-import lombok.Getter;
+import com.github.sarhatabaot.kraken.core.chat.ChatUtil;
+import com.github.sarhatabaot.kraken.core.config.ConfigFile;
 import lombok.Setter;
-import me.tychsen.enchantgui.economy.MoneyPayment;
+import me.tychsen.enchantgui.Main;
 import me.tychsen.enchantgui.economy.NullPayment;
 import me.tychsen.enchantgui.economy.PaymentStrategy;
+import me.tychsen.enchantgui.economy.PlayerPointsPayment;
+import me.tychsen.enchantgui.economy.VaultPayment;
 import me.tychsen.enchantgui.economy.XPPayment;
-import me.tychsen.enchantgui.localization.LocalizationManager;
-import me.tychsen.enchantgui.menu.DefaultMenuSystem;
-import me.tychsen.enchantgui.Main;
-import me.tychsen.enchantgui.util.Common;
 import org.bukkit.command.CommandSender;
-import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.enchantments.Enchantment;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.Map;
 
-/**
- * TODO: make this a static accessor class.
- */
-public class EShopConfig {
-    @Setter
-    @Getter
-    private static EShopConfig instance;
-    @Setter
-    private static FileConfiguration config;
+public class EShopConfig extends ConfigFile<Main> {
+
     @Setter
     private static PaymentStrategy economy;
 
     public EShopConfig() {
-        setInstance(this);
-        setConfig(Main.getInstance().getConfig());
+        super(Main.getInstance(), "", "config.yml", "");
+        saveDefaultConfig();
     }
 
-    public static boolean getIgnoreItemType(){
+
+    public boolean getIgnoreItemType() {
         return config.getBoolean("ignore-itemtype");
     }
 
-    public static boolean getDebug(){
+    public boolean getDebug() {
         return config.getBoolean("debug");
     }
 
-    public static int getPrice(Enchantment enchantment, int level) {
+    public double getPrice(@NotNull Enchantment enchantment, int level) {
         String path = enchantment.getKey().toString().toLowerCase() + ".level" + level;
         path = path.split(":")[1];
-        return config.getInt(path);
+        return config.getDouble(path);
     }
 
-    public static boolean getBoolean(String path){
+    public boolean getBoolean(String path) {
         return config.getBoolean(path);
     }
 
-    public static String getMenuName() {
-        String path = "menu-name";
-        if (config.contains(path) && config.isSet(path) && config.isString(path)) {
-            if (config.getString(path).length() > 32) {
-                return config.getString(path).substring(0, 32);
-            } else {
-                return config.getString(path);
-            }
-        } else {
-            return "EnchantGUI";
+    public String getMenuName() {
+        var path = "menu-name";
+        final String defaultName = "EnchantGUI";
+
+        if (config.getString(path, defaultName).length() > 32) {
+            return config.getString(path, defaultName).substring(0, 32);
         }
+
+        return config.getString(path, defaultName);
+
+
     }
 
-    public static boolean getShowPerItem(){
+    public boolean getShowPerItem() {
         return config.getBoolean("show-per-item");
     }
 
-    public static void reloadConfig(CommandSender sender) {
-        LocalizationManager lm = LocalizationManager.getInstance();
-        if (sender.isOp() || sender.hasPermission("eshop.admin")) {
-            Main.getInstance().reloadConfig();
-            setConfig(Main.getInstance().getConfig());
-            economy = null;
-            Common.tell(sender,DefaultMenuSystem.PREFIX + lm.getString("config-reloaded"));
-        } else {
-            Common.tell(sender,DefaultMenuSystem.PREFIX + lm.getString("no-permission"));
-        }
+    public void reloadConfig(@NotNull CommandSender sender) {
+        reloadConfig();
+        ChatUtil.sendMessage(sender,  Main.getInstance().getLm().getPrefix() + " " + Main.getInstance().getLm().getLanguageString("config-reloaded"));
     }
 
-    public static String[] getEnchantLevels(Enchantment enchantment) {
+
+    public @NotNull String[] getEnchantLevels(@NotNull Enchantment enchantment) {
         String path = enchantment.getKey().toString().toLowerCase();
         path = path.split(":")[1];
         Main.debug(path);
         Map<String, Object> enchantMap = config.getConfigurationSection(path).getValues(false);
         String[] enchantLevels = new String[enchantMap.size()];
 
-        int position = 0;
+        var position = 0;
         for (Map.Entry<String, Object> entry : enchantMap.entrySet()) {
             enchantLevels[position] = entry.getKey();
             position++;
@@ -96,32 +83,32 @@ public class EShopConfig {
         return enchantLevels;
     }
 
-    public static String getEconomyCurrency(){
-        switch (config.getString("payment-currency").toLowerCase()) {
-            case "money":
-                return Main.getEconomy().currencyNameSingular();
-            case "xp":
-                return "levels";
-            default:
-                return "";
-        }
+    public String getEconomyCurrency() {
+        return switch (getPaymentType().toLowerCase()) {
+            case "money" -> Main.getEconomy().currencyNameSingular();
+            case "xp" -> "levels";
+            default -> "";
+        };
     }
 
-    public PaymentStrategy getEconomy() {
+    public String getPaymentType() {
+        return config.getString("payment-currency", "xp");
+    }
+
+    public PaymentStrategy getPaymentStrategy() {
         if (economy == null) {
-            switch (config.getString("payment-currency").toLowerCase()) {
-                case "money":
-                    setEconomy(new MoneyPayment());
-                    break;
-                case "xp":
-                    setEconomy(new XPPayment());
-                    break;
-                default:
-                    setEconomy(new NullPayment());
-                    break;
+            switch (getPaymentType().toLowerCase()) {
+                case "money" -> setEconomy(new VaultPayment());
+                case "xp" -> setEconomy(new XPPayment());
+                case "playerpoints" -> setEconomy(new PlayerPointsPayment());
+                default -> setEconomy(new NullPayment());
             }
         }
         return economy;
+    }
+
+    public String getLanguage() {
+        return config.getString("language", "en");
     }
 
 }
