@@ -1,97 +1,105 @@
 package me.tychsen.enchantgui;
 
-import co.aikar.commands.BukkitCommandManager;
+import co.aikar.commands.PaperCommandManager;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.Setter;
 import me.tychsen.enchantgui.commands.ShopCommand;
 import me.tychsen.enchantgui.config.EShopConfig;
+import me.tychsen.enchantgui.config.Enchants;
 import me.tychsen.enchantgui.event.EventManager;
-import me.tychsen.enchantgui.menu.DefaultMenuSystem;
-import me.tychsen.enchantgui.menu.MenuSystem;
+import me.tychsen.enchantgui.localization.LocalizationManager;
+import me.tychsen.enchantgui.menu.ShopMenu;
 import net.milkbowl.vault.economy.Economy;
+import org.black_ixx.playerpoints.PlayerPoints;
+import org.black_ixx.playerpoints.PlayerPointsAPI;
 import org.bstats.bukkit.Metrics;
-import org.bukkit.event.Listener;
+import org.bukkit.Bukkit;
 import org.bukkit.plugin.RegisteredServiceProvider;
 import org.bukkit.plugin.java.JavaPlugin;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.HashSet;
 import java.util.Set;
 import java.util.UUID;
 
-public class Main extends JavaPlugin implements Listener {
-    @Setter (AccessLevel.PRIVATE)
-    @Getter
+public class Main extends JavaPlugin {
+    @Setter (AccessLevel.PRIVATE) @Getter
     private static Main instance;
-    @Setter (AccessLevel.PRIVATE)
-    @Getter
-    private static Economy economy = null;
-    @Getter
-    @Setter
-    private static MenuSystem menuSystem;
-    @Getter
-    @Setter (AccessLevel.PRIVATE)
+    
+    @Getter @Setter
+    private ShopMenu shopMenu;
+
+    @Getter @Setter (AccessLevel.PRIVATE)
     private static Set<UUID> toggleRightClickPlayers = new HashSet<>();
 
+    private EShopConfig config;
+    private LocalizationManager lm;
+
+    @Getter
+    private PlayerPointsAPI ppApi;
 
     @Override
     public void onEnable() {
         setInstance(this);
-        if (!setupEconomy()) {
-            getLogger().severe("Disabled due to no Vault dependency found!");
-            getServer().getPluginManager().disablePlugin(this);
-            return;
-        }
-        // Generate config.yml if there is none
-        saveDefaultConfig();
+        config = new EShopConfig();
+        lm = new LocalizationManager();
 
+
+        this.shopMenu = new ShopMenu(new Enchants());
         // Register event manager
-        setMenuSystem(new DefaultMenuSystem());
-        getServer().getPluginManager().registerEvents(new EventManager(getMenuSystem()), this);
+        //setMenuSystem(new DefaultMenuSystem());
+        getServer().getPluginManager().registerEvents(new EventManager(),this);
 
         // Register command
-        BukkitCommandManager commandManager = new BukkitCommandManager(this);
+        PaperCommandManager commandManager = new PaperCommandManager(this);
+        commandManager.enableUnstableAPI("brigadier");
         commandManager.registerCommand(new ShopCommand());
 
         // Enable Metrics
         if(!getConfig().getBoolean("opt-out"))
             new Metrics(this, 3871);
 
-        getLogger().info(getName() + " " + getDescription().getVersion() + " enabled!");
+        getLogger().info(() -> getName() + " " + getDescription().getVersion() + " enabled!");
+        getLogger().info(() -> getName() + " " + getDescription().getVersion() + " using: "+ Main.getInstance().getMainConfig().getPaymentStrategy().name());
     }
 
     @Override
     public void onDisable() {
         setInstance(null);
-        setEconomy(null);
-        setMenuSystem(null);
         setToggleRightClickPlayers(null);
-        getLogger().info(getName() + " " + getDescription().getVersion() + " disabled!");
+        getLogger().info(() -> getName() + " " + getDescription().getVersion() + " disabled!");
     }
 
-    public static String getMinecraftVersion(){
-        return instance.getServer().getVersion();
+    public void onReload(){
+        config.reloadConfig();
+
+        hookPlayerPoints();
     }
 
     public static void debug(String msg) {
-        if (EShopConfig.getDebug())
-            Main.getInstance().getLogger().warning(String.format("DEBUG %s", msg));
+        if (Main.getInstance().getMainConfig().getDebug())
+            Main.getInstance().getLogger().info(() -> String.format("DEBUG %s", msg));
     }
 
-    private boolean setupEconomy() {
-        if (getServer().getPluginManager().getPlugin("Vault") == null) {
-            getLogger().severe("could not find vault");
-            return false;
+
+    private void hookPlayerPoints() {
+        if(!config.getPaymentType().equalsIgnoreCase("playerpoints")) {
+            return;
         }
-        RegisteredServiceProvider<Economy> rsp = getServer().getServicesManager().getRegistration(Economy.class);
-        if (rsp == null) {
-            getLogger().severe("could not find Economy.class");
-            return false;
+        if (Bukkit.getPluginManager().isPluginEnabled("PlayerPoints")) {
+            this.ppApi = PlayerPoints.getInstance().getAPI();
         }
-        setEconomy(rsp.getProvider());
-        return economy != null;
     }
 
+    @NotNull
+    public EShopConfig getMainConfig() {
+        return config;
+    }
+
+    public LocalizationManager getLm() {
+        return lm;
+    }
 
 
 }
