@@ -2,62 +2,64 @@ package me.tychsen.enchantgui.commands;
 
 import co.aikar.commands.BaseCommand;
 import co.aikar.commands.annotation.*;
+import com.github.sarhatabaot.kraken.core.chat.ChatUtil;
 import me.tychsen.enchantgui.Main;
-import me.tychsen.enchantgui.config.EShopConfig;
-import me.tychsen.enchantgui.config.EShopShop;
+import me.tychsen.enchantgui.config.Enchants;
 import me.tychsen.enchantgui.localization.LocalizationManager;
-import me.tychsen.enchantgui.menu.MenuSystem;
-import me.tychsen.enchantgui.util.Common;
+import me.tychsen.enchantgui.menu.ShopMenu;
+import me.tychsen.enchantgui.permissions.EShopPermissionSys;
+import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
 @CommandAlias("eshop|enchantgui")
-@CommandPermission("eshop.use")
+@CommandPermission(EShopPermissionSys.USE)
 @Description("Command for the EnchantGUI plugin.")
 public class ShopCommand extends BaseCommand {
-	private String prefix;
-	private MenuSystem menuSystem;
+	private final String prefix;
+	private final ShopMenu shopMenu;
 
-	private LocalizationManager lm = LocalizationManager.getInstance();
+	private final LocalizationManager lm = Main.getInstance().getLm();
 
 	public ShopCommand() {
-		this.prefix = lm.getString("prefix");
-		this.menuSystem = Main.getMenuSystem();
+		this.prefix = lm.getLanguageString("prefix");
+		this.shopMenu = new ShopMenu(new Enchants());
 	}
 
 
 	@Default
 	public void onGui(final Player player){
-		menuSystem.showMainMenu(player);
+		shopMenu.showMainMenu(player);
 	}
 
-	@CommandAlias("toggle")
-	@CommandPermission("eshop.enchantingtable.toggle")
+	@Subcommand("toggle")
+	@CommandPermission(EShopPermissionSys.TOGGLE)
 	public void onToggle(final Player player){
-		if(!EShopConfig.getBoolean("right-click-enchanting-table")){
-			tell(player,lm.getString("disabled-feature"));
+		if(!Main.getInstance().getMainConfig().getBoolean("right-click-enchanting-table")){
+			tell(player,lm.getLanguageString("disabled-feature"));
 			return;
 		}
 
 		if(Main.getToggleRightClickPlayers().contains(player.getUniqueId())){
 			Main.getToggleRightClickPlayers().remove(player.getUniqueId());
-			tell(player,lm.getString("toggle-on"));
+			tell(player,lm.getLanguageString("toggle-on"));
 		}
 		else {
 			Main.getToggleRightClickPlayers().add(player.getUniqueId());
-			tell(player,lm.getString("toggle-off"));
+			tell(player,lm.getLanguageString("toggle-off"));
 		}
 	}
 
 
-	@CommandAlias("reload")
-	@CommandPermission("eshop.reload")
-	public void onReload(final Player player){
-		EShopConfig.reloadConfig(player);
-		LocalizationManager.getInstance().reload(player);
-		EShopShop.getInstance().reload(player);
+	@Subcommand("reload")
+	@CommandPermission(EShopPermissionSys.RELOAD)
+	public void onReload(final CommandSender player){
+		Main.getInstance().onReload();
+		Main.getInstance().getLm().reload(player);
+		this.shopMenu.reload();
+		Main.getInstance().getLogger().info(() -> "%s %s using: %s".formatted(getName(),Main.getInstance().getDescription().getVersion(),Main.getInstance().getMainConfig().getPaymentStrategy().name()));
 	}
 
 	private void tell(final Player player, final String message){
-		Common.tell(player,String.format("%s %s",prefix,message));
+		ChatUtil.sendMessage(player,String.format("%s %s",prefix,message));
 	}
 }
