@@ -1,18 +1,39 @@
 package me.tychsen.enchantgui.economy;
 
 import org.bukkit.entity.Player;
+import org.jetbrains.annotations.NotNull;
 
 /**
  * Used to disable payment.
  */
 public class NullPayment implements PaymentStrategy {
     @Override
-    public boolean withdraw(Player p, int amount) {
+    public String name() {
+        return "NullPayment";
+    }
+
+    @Override
+    public boolean withdraw(@NotNull Player player, int amount) {
         return true;
     }
 
     @Override
-    public boolean hasSufficientFunds(Player p, int amount) {
+    public boolean hasSufficientFunds(@NotNull Player player, int amount) {
         return true;
+    }
+
+    @Override
+    public boolean withdraw(@NotNull final Player player, final double amount) {
+        return true;
+    }
+
+    @Override
+    public boolean hasSufficientFunds(@NotNull final Player player, final double amount) {
+        return true;
+    }
+
+    @Override
+    public String getCurrency() {
+        return "";
     }
 }
