@@ -1,46 +1,57 @@
 package me.tychsen.enchantgui.permissions;
 
-import lombok.RequiredArgsConstructor;
 import org.bukkit.enchantments.Enchantment;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
+import org.jetbrains.annotations.Contract;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.Map;
 
 public class EShopPermissionSys {
-    public boolean hasEnchantPermission(Player p, ItemStack item) {
-        if (p.isOp()) return true;
+    public static final String USE = "eshop.use";
+    public static final String ENCHANTING_TABLE = "eshop.enchantingtable";
+    public static final String TOGGLE = "eshop.enchantingtable.toggle";
+    public static final String RELOAD = "eshop.reload";
+
+    private static final String BASE = "eshop.enchants.";
+
+    public static boolean hasEnchantPermission(final @NotNull Player player, final ItemStack item) {
+        if (player.isOp()) return true;
 
         Map<Enchantment, Integer> enchants = item.getEnchantments();
-        if (enchants.size() > 1) throw new TooManyEnchantmentsException("Item has more than one enchant!");
+        if (enchants.size() > 1) {
+            throw new TooManyEnchantmentsException("Item has more than one enchant!");
+            //This should never happen.
+        }
 
-        Enchantment ench = enchants.keySet().toArray(new Enchantment[1])[0];
-        String base = "eshop.enchants.";
-        String name = (ench.getKey().toString().toLowerCase()).split(":")[1];
-        String perm = base + name;
-
-        return p.hasPermission(perm) || p.hasPermission(base + "all");
+        Enchantment enchantment = enchants.keySet().toArray(new Enchantment[1])[0];
+        String name = (enchantment.getKey().toString().toLowerCase()).split(":")[1];
+        String perm = getEnchantmentPermission(name);
+        return player.hasPermission(perm) || player.hasPermission(BASE + "all");
     }
 
-    public boolean hasEnchantPermission(Player p, Enchantment ench, int level) {
-        if (p.isOp()) return true;
+    public static boolean hasEnchantPermission(final @NotNull Player player, final Enchantment enchantment, int level) {
+        if (player.isOp()) return true;
+        String enchantmentName = (enchantment.getKey().toString().toLowerCase()).split(":")[1];
+        String perm = getEnchantmentLevelPermission(enchantmentName,level);
 
-        String base = "eshop.enchants.";
-        String enchName = (ench.getKey().toString().toLowerCase()).split(":")[1];
-        String perm = base + enchName + "." + level;
-
-        return p.hasPermission(perm) || p.hasPermission(base + enchName + ".all") || p.hasPermission(base + "all");
+        return player.hasPermission(perm) || player.hasPermission(BASE + enchantmentName + ".all") || player.hasPermission(BASE + "all");
     }
 
-    public boolean hasUsePermission(Player p) {
-        if (p.isOp()) return true;
-        return  p.hasPermission("eshop.use");
+    public static boolean hasUsePermission(@NotNull Player player) {
+        if (player.isOp()) return true;
+        return player.hasPermission(EShopPermissionSys.USE);
     }
 
-    @RequiredArgsConstructor
-    private static final class TooManyEnchantmentsException extends RuntimeException {
-        private static final long serialVersionUID = 1L;
-
-        private final String tellMessage;
+    @Contract(pure = true)
+    private static @NotNull String getEnchantmentPermission(final String name) {
+        return BASE + name;
     }
+
+    @Contract(pure = true)
+    private static @NotNull String getEnchantmentLevelPermission(final String name, int level) {
+        return getEnchantmentPermission(name) + "."+level;
+    }
+
 }
